requirements.txt
aniso8601==2.0.0
astroid==1.6.1
attrs==17.4.0
blinker==1.4
click==6.7
Faker==0.8.10
Flask==0.12.2
Flask-Cors==3.0.3
Flask-JSON==0.3.2
Flask-JWT-Extended==3.7.0
Flask-Mail==0.9.1
flask-mongoengine==0.9.3
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-WTF==0.14.2
freezegun==0.3.9
isort==4.3.3
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
mongoengine==0.15.0
mongomock==3.8.0
parso==0.1.1
pbr==3.1.1
pep8==1.7.1
pluggy==0.6.0
py==1.5.2
PyJWT==1.5.3
pylint==1.8.2
pylint-django==0.9.0
pylint-plugin-utils==0.2.6
pymongo==3.6.0
pytest==3.4.0
python-dateutil==2.6.1
pytz==2017.3
sentinels==1.0.0
six==1.11.0
text-unidecode==1.1
uWSGI==2.0.15
uwsgidecorators==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
